// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const FeedbackType = {
  BUG: 'BUG',
  FEATURE: 'FEATURE',
  GENERAL: 'GENERAL'
} as const

export type FeedbackType = (typeof FeedbackType)[keyof typeof FeedbackType]